+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Add option to install with homebrew
+- Added `condition` option to tasks
 
 ## v1.0.0 - 2023-05-23